requirements_improved.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database
pymongo==4.6.0
python-dotenv==1.0.0

# AI and ML
openai==1.3.7
youtube-transcript-api==1.2.1
yt-dlp==2023.11.16

# HTTP requests
requests==2.31.0

# Async support
asyncio==3.4.3

# File processing
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.7

# Logging and monitoring
structlog==23.2.0

# Type hints
typing-extensions==4.8.0 